refactor: move health score contracts to output crate
